Add ruby-path package version 1.3.1.
Path - a Path manipulation library

Path is a library to manage paths.  It is similar to Pathname, but has some
extra goodness.  The method names are intended to be short and explicit, and
avoid too much duplication like having 'name' or 'path' in the method name.

I believe the object-oriented approach to manipulate paths is very elegant and
useful.  Paths are naturally the subject of their methods and even if they are
simple Strings behind, they carry way much more information and deserve a
first-class status.

Also, using a path library like this avoid to remember in which class the
functionality is implemented, everything is in one place (if not, please open
an issue!).