Add p5-Test-File.
This module provides a collection of test utilities for file attributes.

Some file attributes depend on the owner of the process testing the
file in the same way the file test operators do. For instance, root
(or super-user or Administrator) may always be able to read files
no matter the permissions.

Some attributes don't make sense outside of Unix, either, so some
tests automatically skip if they think they won't work on the
platform. If you have a way to make these functions work on Windows,
for instance, please send me a patch. :)