Update to Asterisk 11.3.0:  this is a bugfix release.
The Asterisk Development Team has announced the release of Asterisk 11.3.0.

The release of Asterisk 11.3.0 resolves several issues reported by the
community and would have not been possible without your participation.
Thank you!

The following is a sample of the issues resolved in this release:

* --- Fix issue where chan_mobile fails to bind to first available port

* --- Fix Queue Log Reporting Every Call COMPLETECALLER With "h"
      Extension Present

* --- Retain XMPP filters across reconnections so external modules
      continue to function as expected.

* --- Ensure that a declined media stream is terminated with a '\r\n'

* --- Fix pjproject compilation in certain circumstances

For a full list of changes in this release, please see the ChangeLog:

http://downloads.asterisk.org/pub/telephony/asterisk/ChangeLog-11.3.0

Thank you for your continued support of Asterisk!