Import ttf-tlwg-0.6.0 as fonts/ttf-tlwg.
Fonts-TLWG (formerly ThaiFonts-Scalable) is a collection of Thai
scalable fonts available in free licenses. Its goal is to provide
fonts that conform to existing standards and recommendations, so
that it can be a reference implementation.

So, the fonts have to be modified from their original versions as
needed, according to following references:

* Recommendation for Thai font design, from the National Font Project
* Adobe Glyph naming convention
* Microsoft Typography
* Common practice for Thai Windows fonts